Improve documentation on weather API and general code refactoring
The documentation was improved on endpoint and data validation schemas of the weather.ts file. The doc-strings were added to improve the understanding on the purpose of these structures and major functions such as how to calculate the Air Quality Index (AQI). The refactoring included removing the unnecessary comma at the end of the object definitions, which enhances the consistency and readability of the codes and deleted console.log used for debugging that is no longer needed. It was further validated that precipitation probabilities should only be included in the response when hourly data is available.