Moon Phase - Bug Fix
I check if the prop (translation of the word)
is one word or more. Have the translation more than one word than I use break words.
Otherwise I use break all.
The next Problem was, that I thought I musst optimize
it for Safari.
But the Problem is Webkit. I musst optimize it for all Browsers that the user can open on a IPhone.